# apps/backend/api-docs.yaml
openapi: 3.0.3
info:
  title: Meridian Backend API
  description: |
    Meridian是一个AI驱动的个性化情报简报系统API。
    
    该API提供以下主要功能：
    - RSS源管理和文章抓取
    - AI驱动的文章分析和处理
    - 智能简报生成
    - 系统监控和可观测性
    - Durable Objects分布式管理
    
    ## 认证
    
    大部分管理端点需要Bearer Token认证。在请求头中包含：
    ```
    Authorization: Bearer {API_TOKEN}
    ```
    
    ## 响应格式
    
    API使用统一的响应格式：
    ```json
    {
      "success": boolean,
      "data": any,
      "message": string,
      "timestamp": string,
      "pagination": object  // 可选，仅分页查询
    }
    ```
    
    ## 错误处理
    
    错误响应格式：
    ```json
    {
      "success": false,
      "error": string,
      "timestamp": string
    }
    ```
  version: 1.0.0
  contact:
    name: Meridian API Support
    url: https://github.com/your-org/meridian
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8787
    description: 开发环境
  - url: https://api.meridian.example.com
    description: 生产环境
security:
  - bearerAuth: []
paths:
...
  /admin/sources:
    get:
      summary: 获取RSS源列表
      description: 获取所有配置的RSS源
      tags:
        - RSS源管理
      responses:
        '200':
          description: 成功获取RSS源列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  - id: 1
                    url: "https://news.ycombinator.com/rss"
                    name: "Hacker News"
                    category: "tech"
                    scrape_frequency: 4
                    paywall: false
                    lastChecked: "2025-01-01T12:00:00Z"
                message: "获取了3个RSS源"
                timestamp: "2025-01-01T12:00:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: 创建新的RSS源
      description: 添加一个新的RSS源到系统
      tags:
        - RSS源管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        '201':
          description: RSS源创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  id: 4
                  url: "https://example.com/feed.xml"
                  name: "Example Feed"
                  category: "news"
                  scrape_frequency: 4
                message: "RSS源添加成功"
                timestamp: "2025-01-01T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/sources/{id}:
    put:
      summary: 更新RSS源
      description: 更新指定RSS源的信息
      tags:
        - RSS源管理
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: RSS源ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSourceRequest'
      responses:
        '200':
          description: RSS源更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: 删除RSS源
      description: 删除指定的RSS源（基础删除，不包含DO清理）
      tags:
        - RSS源管理
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: RSS源ID
      responses:
        '200':
          description: RSS源删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/articles:
    get:
      summary: 获取文章列表
      description: 获取文章列表，支持分页和状态过滤
      tags:
        - 文章管理
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ArticleStatus'
          description: 文章状态过滤
      responses:
        '200':
          description: 成功获取文章列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  - id: 452
                    title: "Twenty years strong: a love letter to TechCrunch"
                    url: "https://techcrunch.com/example-article"
                    status: "PROCESSED"
                    publishDate: "2025-01-01T10:00:00Z"
                    processedAt: "2025-01-01T11:00:00Z"
                    sourceId: 2
                    contentFileKey: "articles/452.txt"
                pagination:
                  page: 1
                  limit: 20
                  total: 209
                message: "获取了20篇文章"
                timestamp: "2025-01-01T12:00:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/overview:
    get:
      summary: 系统概览
      description: 获取系统运行状态的概览信息
      tags:
        - 系统管理
      responses:
        '200':
          description: 成功获取系统概览
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  sources:
                    total: 3
                    active: 3
                  articles:
                    total: 209
                    processed: 141
                    pending: 50
                    failed: 18
                  briefs:
                    last30Days: 16
                  lastUpdated: "2025-01-01T12:00:00Z"
                message: "系统概览获取成功"
                timestamp: "2025-01-01T12:00:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/briefs/generate:
    post:
      summary: 生成简报
      description: 触发简报生成工作流
      tags:
        - 简报管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateBriefRequest'
      responses:
        '202':
          description: 简报生成工作流已启动
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  workflowId: "workflow_12345"
                message: "简报生成工作流已启动"
                timestamp: "2025-01-01T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # 简报端点 (/reports)
  # ============================================================================
...
  /sources/{id}:
    delete:
      summary: 删除RSS源（高级）
      description: 删除RSS源并清理相关的Durable Objects
      tags:
        - 高级源管理
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: RSS源ID
      responses:
        '200':
          description: RSS源删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

# ============================================================================
# 组件定义
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "使用API Token进行认证，格式：Bearer {API_TOKEN}"
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        data:
          description: 响应数据
        message:
          type: string
          description: 响应消息
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
        - success
        - timestamp

    # 分页信息
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: 当前页码
        limit:
          type: integer
          description: 每页数量
        total:
          type: integer
          description: 总记录数
        hasMore:
          type: boolean
          description: 是否还有更多数据

    # RSS源相关
...
    CreateSourceRequest:
      type: object
      properties:
        name:
          type: string
          description: 源名称
          minLength: 1
        url:
          type: string
          format: uri
          description: RSS源URL
        category:
          type: string
          nullable: true
          description: 分类
        scrape_frequency:
          type: integer
          description: 抓取频率
          minimum: 1
          maximum: 4
          default: 4
      required:
        - name
        - url
    UpdateSourceRequest:
      type: object
      properties:
        name:
          type: string
          description: 源名称
        url:
          type: string
          format: uri
          description: RSS源URL
        category:
          type: string
          nullable: true
          description: 分类
        scrape_frequency:
          type: integer
          description: 抓取频率
          minimum: 1
          maximum: 4

    # 文章相关
    ArticleStatus:
      type: string
      enum:
        - PENDING_FETCH
        - CONTENT_FETCHED
        - PROCESSED
        - SKIPPED_PDF
        - FETCH_FAILED
        - RENDER_FAILED
        - AI_ANALYSIS_FAILED
        - EMBEDDING_FAILED
        - R2_UPLOAD_FAILED
        - SKIPPED_TOO_OLD
      description: 文章处理状态
...
    GenerateBriefRequest:
      type: object
      properties:
        dateFrom:
          type: string
          format: date-time
          description: 开始日期
        dateTo:
          type: string
          format: date-time
          description: 结束日期
        minImportance:
          type: integer
          description: 最小重要性分数
          minimum: 1
          maximum: 10
          default: 5
        triggeredBy:
          type: string
          description: 触发方式
          default: "manual"
      required:
        - dateFrom
        - dateTo

    # 可观测性相关
...
  responses:
...
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "缺少必需的字段: name, url"
              timestamp:
                type: string
                format: date-time
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
    NotFound:
      description: 资源未找到
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "未找到指定的RSS源"
              timestamp:
                type: string
                format: date-time
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "该URL已存在"
              timestamp:
                type: string
                format: date-time
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Internal server error"
              timestamp:
                type: string
                format: date-time
...